+++ biblio/bibliography.bib
@@ -90,6 +90,16 @@
 % C
 %-----------------------------
 
+@misc{koyama2026computationdegree1persistenthomology,
+  title         = {Computation of degree-1 persistent homology on larger point-clouds using the Reduced Vietoris-Rips filtration},
+  author        = {Musashi Ayrton Koyama and Facundo M{\'e}moli and Vanessa Robins and Katharine Turner},
+  year          = {2026},
+  eprint        = {2307.16333},
+  archiveprefix = {arXiv},
+  primaryclass  = {math.AT},
+  url           = {https://arxiv.org/abs/2307.16333}
+}
+
 @book{kaczynski2004computational,
   title     = {Computational Homology},
   author    = {Kaczynski, T. and Mischaikow, K. and Mrozek, M.},
@@ -434,15 +444,6 @@
 % K
 %-----------------------------
 
-@misc{koyama2026computationdegree1persistenthomology,
-  title         = {Computation of degree-1 persistent homology on larger point-clouds using the Reduced Vietoris-Rips filtration},
-  author        = {Musashi Ayrton Koyama and Facundo M{\'e}moli and Vanessa Robins and Katharine Turner},
-  year          = {2026},
-  eprint        = {2307.16333},
-  archiveprefix = {arXiv},
-  primaryclass  = {math.AT},
-  url           = {https://arxiv.org/abs/2307.16333}
-}
 
 %-----------------------------
 % L
