+++ src/Reduced_rips/include/gudhi/Reduced_rips/Euclidean_kd_tree.h
@@ -11,55 +11,53 @@
  *    - YYYY/MM Author: Description of the modification
  */
 
+/**
+ * @file Euclidean_kd_tree.h
+ * @author Thomas Burnett, Musashi Koyama
+ * @brief Spatial neighbor search over a Euclidean point cloud: a CGAL kd-tree in low ambient dimension, a
+ * flat brute-force scan otherwise.
+ */
+
 #ifndef REDUCED_RIPS_EUCLIDEAN_KD_TREE_H_
 #define REDUCED_RIPS_EUCLIDEAN_KD_TREE_H_
 
 #include <algorithm>
 #include <cstddef>
 #include <iterator>
-#include <memory>
-#include <numeric>
+#include <optional>
 #include <utility>
 #include <vector>
 
 #include <CGAL/Epick_d.h>
 
 #include <gudhi/Kd_tree_search.h>
-#include <gudhi/Reduced_rips/Cloud.h>
+#include <gudhi/Reduced_rips/Helpers.h>
 
 namespace Gudhi {
 
 namespace reduced_rips {
 
-using detail::Cloud;
-using detail::l2_dist_2;
-
 // Spatial search over the cloud. In dimension <= 3 it uses GUDHI's Kd_tree_search. In dimension >= 4,
 // it falls back to a flat brute-force scan over the Cloud. Every distance comparison uses l2_dist_2.
 class Euclidean_kd_tree {
+  using Cloud = detail::Cloud;
+
  public:
-  Euclidean_kd_tree(const Cloud& pm, bool brute_force) : pm_(pm), brute_force_(brute_force) {
-    if (brute_force_) return;
+  Euclidean_kd_tree(const Cloud& pm, bool brute_force) : pm_(pm) {
+    if (brute_force) return;  // leave tree_ disengaged; queries then fall back to a brute-force scan
     kd_points_.reserve(pm.n);
     for (std::size_t i = 0; i < pm.n; ++i) kd_points_.emplace_back(pm[i], pm[i] + pm.dim);
     // Kd_tree_search keeps a reference to kd_points_, which outlives the tree (both are members below).
-    tree_ = std::make_unique<Kd_tree>(kd_points_);
+    tree_.emplace(kd_points_);
   }
 
   // Indices of the k nearest points to query, ascending by squared distance (ties by index). A query that
   // is itself a Cloud point is returned as the nearest (distance 0); callers filter that out.
   std::vector<std::size_t> nearest_neighbors(const double* query, std::size_t k) const {
-    if (brute_force_) {
+    if (!tree_) {
       std::vector<double> dist(pm_.n);
-      for (std::size_t i = 0; i < pm_.n; ++i) dist[i] = l2_dist_2<double>(pm_[i], query, pm_.dim);
-      k = std::min(k, pm_.n);
-      std::vector<std::size_t> result(pm_.n);
-      std::iota(result.begin(), result.end(), std::size_t{0});
-      std::partial_sort(
-          result.begin(), result.begin() + static_cast<std::ptrdiff_t>(k), result.end(),
-          [&dist](std::size_t x, std::size_t y) { return dist[x] != dist[y] ? dist[x] < dist[y] : x < y; });
-      result.resize(k);
-      return result;
+      for (std::size_t i = 0; i < pm_.n; ++i) dist[i] = detail::l2_dist_2<double>(pm_[i], query, pm_.dim);
+      return detail::smallest_indices_by(0, pm_.n, std::min(k, pm_.n), [&dist](std::size_t x) { return dist[x]; });
     }
     Kd_point center(query, query + pm_.dim);
     std::vector<std::size_t> result;
@@ -69,24 +67,23 @@ class Euclidean_kd_tree {
   }
 
   // All points within the given squared radius of query, as (index, squared distance) pairs. The radius and the
-  // returned squared distances are in the scalar type T; the CGAL tree searches in double, so the radius is
-  // narrowed to double for the query and each survivor's distance is recomputed in T afterwards.
+  // returned squared distances are in the scalar type T; the CGAL tree searches in double, and each survivor's
+  // distance is recomputed in T afterwards.
   template <class T>
   std::vector<std::pair<std::size_t, T>> points_in_squared_ball(const double* query, T squared_radius) const {
     std::vector<std::pair<std::size_t, T>> result;
-    if (brute_force_) {
+    if (!tree_) {
       for (std::size_t i = 0; i < pm_.n; ++i) {
-        T d = l2_dist_2<T>(pm_[i], query, pm_.dim);
+        T d = detail::l2_dist_2<T>(pm_[i], query, pm_.dim);
         if (d <= squared_radius) result.emplace_back(i, d);
       }
       return result;
     }
     Kd_point center(query, query + pm_.dim);
     std::vector<std::size_t> found;
-    const auto radius = static_cast<double>(squared_radius);
-    tree_->all_near_neighbors2(center, radius, radius, std::back_inserter(found));
+    tree_->all_near_neighbors2(center, squared_radius, squared_radius, std::back_inserter(found));
     result.reserve(found.size());
-    for (std::size_t idx : found) result.emplace_back(idx, l2_dist_2<T>(pm_[idx], query, pm_.dim));
+    for (std::size_t idx : found) result.emplace_back(idx, detail::l2_dist_2<T>(pm_[idx], query, pm_.dim));
     return result;
   }
 
@@ -96,9 +93,8 @@ class Euclidean_kd_tree {
   using Kd_tree = Gudhi::spatial_searching::Kd_tree_search<Kd_kernel, std::vector<Kd_point>>;
 
   Cloud pm_;  // non-owning view (pointer + sizes); pointee outlives this
-  bool brute_force_;
   std::vector<Kd_point> kd_points_;
-  std::unique_ptr<Kd_tree> tree_;
+  std::optional<Kd_tree> tree_;  // disengaged in brute-force mode; engaged holds the kd-tree over kd_points_
 };
 
 }  // namespace reduced_rips
